+++ src/main/java/com/epam/reportportal/utils/properties/ListenerProperty.java
@@ -151,9 +151,14 @@ public enum ListenerProperty {
     ATTRIBUTE_NUMBER_LIMIT("rp.attribute.limit", false),
 
     /**
-     * Enable/Disable the feature to truncate Stack Traces of exceptions that being logged to ReportPortal. Default value: <code>true</code>.
+     * Toggle the feature to truncate Stack Traces of exceptions that being logged to ReportPortal. Default value: <code>true</code>.
      */
     EXCEPTION_TRUNCATE("rp.truncation.exception", false),
+
+	/**
+	 * Toggle the feature to replace basic binary characters with \uFFFD char. Default value: <code>true</code>.
+	 */
+	REPLACE_BINARY_CHARACTERS("rp.sanitization.replace.binary", false),
 
     // Issue reporting properties
     /**