+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Attribute number truncation, by @HardNorth
+- Binary character replacement in basic text fields, by @HardNorth
 
 ## [5.4.11]
 ### Changed